move rewards to own file
